Upgraded api, added docker support
