--- api/.dockerignore
@@ -1,14 +1,0 @@
-.cenv/
-
-*.pyc
-__pycache__/
-
-instance/
-
-.pytest_cache/
-.coverage
-htmlcov/
-
-dist/
-build/
-*.egg-info/
