+++ api/.gitignore
@@ -1,14 +1,4 @@
-.cenv/
-
-*.pyc
-__pycache__/
-
-instance/
-
-.pytest_cache/
-.coverage
-htmlcov/
-
 dist/
 build/
+Data/
 *.egg-info/
