+++ api/flaskr/__init__.py
@@ -1,5 +1,17 @@
-from .app import create_app
-from dotenv import load_dotenv
-load_dotenv()
+from flask import Flask
+import os
 
-app = create_app()
+UPLOAD_FOLDER = 'Data'
+ALLOWED_EXTENSIONS = {'csv', 'xlsx'}
+
+app = Flask(__name__, template_folder='templates')
+
+application = app
+application.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
+
+upload_folder = application.config['UPLOAD_FOLDER']
+if not os.path.exists(upload_folder):
+    os.makedirs(upload_folder)
+
+from . import api
+
