--- api/flaskr/api/__init__.py
@@ -1,5 +1,0 @@
-from flask import Blueprint, request
-
-api = Blueprint("api", __name__)
-
-from . import apiController
