+++ api/flaskr/static/style/styles.css
@@ -1,0 +1,226 @@
+* {
+    box-sizing: border-box;
+    margin: 0;
+    padding: 0;
+}
+
+.index_background {
+    position: relative;
+    background: url("../style/img/back2.jpg");
+    box-sizing: 100%;
+    background-repeat: no-repeat;
+    background-position: center;
+    background-size: cover;
+    border: none;
+    background-color: rgba(34, 33, 32, 0.5);
+    background-blend-mode: multiply;
+}
+
+.main_header {
+    display: flex;
+    flex-direction: row;
+    background-color: #820d0d;
+    color: white;
+    padding: 10px 0;
+    font-size: 20px;
+    font-family: "Ubuntu", sans-serif;
+}
+.main_header nav {
+    display: flex;
+    justify-content: center;
+}
+.main_header nav a {
+    color: white;
+    padding: 14px 20px;
+    text-decoration: none;
+    text-align: center;
+    font-family: "Roboto", sans-serif;
+    font-weight: 200;
+    font-style: normal;
+}
+.main_header nav a:hover {
+    background-color: #ef6868;
+    border-radius: 20px;
+}
+
+.content {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    color: rgb(218, 194, 194);
+    font-size: 30px;
+    font-family: "Ubuntu", sans-serif;
+    margin-top: 130px;
+    font-size: 30px;
+}
+
+.footer_h {
+    background-color: #292525;
+    font-family: "Ubuntu", sans-serif;
+    padding: 10px;
+    color: aliceblue;
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    width: 100%;
+    height: 80px;
+    margin-top: 140px;
+}
+
+.form {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 40px;
+    margin-top: 50px;
+    width: 500px;
+    height: 400px;
+    border-radius: 50px;
+    background-color: #820d0d;
+    opacity: .75;
+}
+.form_value {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+}
+
+.form_label {
+    padding: 0px 20px 20px 0px;
+    font-size: 35px;
+}
+
+.form_button {
+    width: 420px;
+    height: 30px;
+    border-radius: 20px;
+    font-size: 20px;
+    background-color: #e47575;
+    border: none;
+    color: #820d0d;
+}
+
+.input_form {
+    width: 450px;
+    height: 230px;
+    font-size: 22px;
+}   
+
+.main_bars {
+    padding-top: 60px;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    font-size: 30px;
+    margin-left: 120px;
+}
+.main_bars h1 {
+    margin-top: 70px;
+    display: flex;
+    font-size: 35px;
+    font-family: "Ubuntu", sans-serif;
+    flex-direction: column;
+    align-items: center;
+    font-family: "Roboto", sans-serif;
+    color: #820d0d;
+}
+
+.main_inf {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    width: 540px;
+    padding: 20px;
+}
+
+.rating {
+    background-color: #fff0f0;
+}
+
+.first_buttons{
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+}
+
+.first_but {
+    font-size: 15px;
+    font-family: "Ubuntu", sans-serif;
+    margin-right: 30px;
+    width: 200px;
+    height: 40px;
+    border-radius: 10px;
+    background-color: #fff0f0;
+    border: none;
+    color: #820d0d;
+}
+
+.first_but:hover {
+    background-color: #fa9b9b;
+}
+
+.info_buttons {
+    display: flex;
+    flex-direction: row;
+}
+
+.info_but {
+    margin: 20px;
+    font-size: 15px;
+    font-family: "Ubuntu", sans-serif;
+    margin-right: 30px;
+    width: 200px;
+    height: 40px;
+    border-radius: 30px;
+    background-color: #ffc5c5;
+    border: none;
+    color: #820d0d;
+}
+
+.h2_details {
+    display: flex;
+    justify-content: center;
+    font-size: 20px;
+    font-family: "Ubuntu", sans-serif;
+    color: #820d0d;
+}
+
+.h2_buttons {
+    display: flex;
+    justify-content: center;
+    font-size: 12px;
+    margin: 20px;
+    font-family: "Ubuntu", sans-serif;
+    color: #fff0f0;
+}
+
+.rating_footer {
+    background-color: #292525;
+    font-family: "Ubuntu", sans-serif;
+    color: aliceblue;
+    display: flex;
+    flex-direction: column;
+    margin-top: 50px;
+    padding-top: 10px;
+    padding-bottom: 30px;
+}
+
+.chart-container {
+    display: flex;
+    width: 450px;
+    padding: 40px;
+}
+
+.chart-row{
+    display: flex;
+    flex-direction: row;
+    padding: 20px;
+}
+
+.dop_raiting_bars {
+    display: flex;
+    flex-direction: row;
+}
