+++ api/flaskr/templates/dop_rating.html
@@ -1,0 +1,206 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>RZD</title>
+    <!-- <script src="https://cdn.jsdelivr.net/npm/chart.js"></script> -->
+    <link rel="stylesheet" href="\static\style\styles.css">
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Pacifico&family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&family=Ubuntu:ital,wght@0,300;0,400;0,500;0,700;1,300;1,400;1,500;1,700&display=swap" rel="stylesheet">
+</head>
+<body>
+    
+    <section class="rating">
+        <header class="main_header">
+            <nav>
+                <!-- <img src="https://polotnos.cdnbro.com/posts/859707-rzhd-logotip-24.jpg  " alt="" style="width: 100px;"> -->
+                <a href="/">Загрузить новый файл</a>
+                <a href="/rating">Общий отчёт</a>
+            </nav>
+        </header>
+        <div class="main_bars">
+            <h1>Рейтинг полигонов по штрафам, манере вождения и путям</h1>
+                <!-- Вторая строка графиков -->
+                <div class="dop_raiting_bars">
+                    <div class="chart-container">
+                        <canvas id="sub1_raiting" width="120" height="120"></canvas>
+                    </div>
+                    <div class="chart-container">
+                        <canvas id="sub2_raiting" width="120" height="120"></canvas>
+                    </div>
+                    <div class="chart-container">
+                        <canvas id="sub3_raiting" width="120" height="120"></canvas>
+                    </div>
+                </div>
+            </div>
+
+        </div>
+        <footer class="rating_footer">
+            <div class="h2_buttons">
+                <h2>Детальная информация о каждом полигоне</h2>
+            </div>
+            <div class="first_buttons">
+                <button class="first_but">Московская ЖД</button>
+                <button class="first_but">Калининградская ЖД</button>
+                <button class="first_but">Октябрьская ЖД</button>
+                <button class="first_but">Горьковская ЖД</button>
+                <button class="first_but">Северная ЖД</button>
+            </div>
+        </footer>
+    </section>
+    <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
+    <script>
+        var labels1 = {{ labels_mileage_deviation_score | tojson }};
+        var values1 = {{ values_mileage_deviation_score | tojson }};
+
+        // Создаем массив объектов для сортировки
+        var data = labels1.map((label, index) => ({
+            label: label,
+            value: values1[index]
+        }));
+
+        // Сортируем данные по значению, перемещая NaN в конец
+        data.sort((a, b) => {
+            if (isNaN(a.value) && isNaN(b.value)) {
+                return 0;
+            } else if (isNaN(a.value)) {
+                return 1;
+            } else if (isNaN(b.value)) {
+                return -1;
+            } else {
+                return b.value - a.value;
+            }
+        });
+
+        // Разделяем данные обратно на массивы labels и values
+        labels1 = data.map(item => item.label);
+        values1 = data.map(item => item.value);
+
+        var ctx1 = document.getElementById('sub1_raiting');
+        new Chart(ctx1, {
+            type: 'bar',
+            data: {
+                labels: labels1,
+                datasets: [{
+                    label: 'Рейтинг полигонов по пробегу',
+                    data: values1,
+                    backgroundColor: 'rgba(75, 192, 192)',
+                    borderColor: 'rgba(75, 192, 192)',
+                    borderWidth: 1
+                }]
+            },
+            options: {
+                scales: {
+                    yAxes: [{
+                        ticks: {
+                            beginAtZero: true
+                        }
+                    }]
+                }
+            }
+        });
+
+        var labels2 = {{ labels_driving_style_score | tojson }};
+        var values2 = {{ values_driving_style_score | tojson }};
+
+        // Создаем массив объектов для сортировки
+        var data = labels2.map((label, index) => ({
+            label: label,
+            value: values2[index]
+        }));
+
+        // Сортируем данные по значению, перемещая NaN в конец
+        data.sort((a, b) => {
+            if (isNaN(a.value) && isNaN(b.value)) {
+                return 0;
+            } else if (isNaN(a.value)) {
+                return 1;
+            } else if (isNaN(b.value)) {
+                return -1;
+            } else {
+                return b.value - a.value;
+            }
+        });
+
+        // Разделяем данные обратно на массивы labels и values
+        labels2 = data.map(item => item.label);
+        values2 = data.map(item => item.value);
+
+        var ctx2 = document.getElementById('sub2_raiting');
+        new Chart(ctx2, {
+            type: 'bar',
+            data: {
+                labels: labels2,
+                datasets: [{
+                    label: 'Рейтинг полигонов по стилю езды',
+                    data: values2,
+                    backgroundColor: 'rgba(75, 192, 192)',
+                    borderColor: 'rgba(75, 192, 192)',
+                    borderWidth: 1
+                }]
+            },
+            options: {
+                scales: {
+                    yAxes: [{
+                        ticks: {
+                            beginAtZero: true
+                        }
+                    }]
+                }
+            }
+        });
+
+        var labels3 = {{ labels_penalty_score | tojson }};
+        var values3 = {{ values_penalty_score | tojson }};
+
+        // Создаем массив объектов для сортировки
+        var data = labels3.map((label, index) => ({
+            label: label,
+            value: values3[index]
+        }));
+
+        // Сортируем данные по значению, перемещая NaN в конец
+        data.sort((a, b) => {
+            if (isNaN(a.value) && isNaN(b.value)) {
+                return 0;
+            } else if (isNaN(a.value)) {
+                return 1;
+            } else if (isNaN(b.value)) {
+                return -1;
+            } else {
+                return b.value - a.value;
+            }
+        });
+
+        // Разделяем данные обратно на массивы labels и values
+        labels3 = data.map(item => item.label);
+        values3 = data.map(item => item.value);
+
+        var ctx3 = document.getElementById('sub3_raiting');
+        new Chart(ctx3, {
+            type: 'bar',
+            data: {
+                labels: labels3,
+                datasets: [{
+                    label: 'Рейтинг полигонов по штрафам',
+                    data: values3,
+                    backgroundColor: 'rgba(75, 192, 192)',
+                    borderColor: 'rgba(75, 192, 192)',
+                    borderWidth: 1
+                }]
+            },
+            options: {
+                scales: {
+                    yAxes: [{
+                        ticks: {
+                            beginAtZero: true
+                        }
+                    }]
+                }
+            }
+        });
+    </script>
+</body>
+</html>
