+++ api/flaskr/templates/index.html
@@ -1,0 +1,35 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>RZD</title>
+    <link rel="stylesheet" href="\static\style\styles.css">
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Pacifico&family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&family=Ubuntu:ital,wght@0,300;0,400;0,500;0,700;1,300;1,400;1,500;1,700&display=swap" rel="stylesheet">
+</head>
+<body class="index_background">
+    <header class="main_header">
+        <nav>
+            <!-- <img src="https://polotnos.cdnbro.com/posts/859707-rzhd-logotip-24.jpg  " alt="" style="width: 100px;"> -->
+            <a href="/">Загрузить файл</a>
+            <a href="/rating">Вернуться к прошлому файлу</a>
+        </nav>
+    </header>
+    <div id="home" class="content">
+        <h1>Добро пожаловать, загрузите файл!</h1>
+        <form action="/setData" method="post" enctype="multipart/form-data" class="form">
+            <div class="form_value">
+                <label for="fileUpload" class="form_label">Загрузить файл:</label>
+                <input type="file" id="fileUpload" name="file" class="input_form">
+                <button type="submit" class="form_button">Отправить данные</button>
+            </div>
+        </form>
+    </div>
+    <footer>
+        <h2 class="footer_h">Made by Clio</h2>
+    </footer>
+    <script src="\static\script\scripts.js"></script>
+</body>
+</html>
