--- api/pyproject.toml
@@ -1,15 +1,0 @@
-[project]
-name="flaskr"
-version="0.0.1"
-dynamic = ["dependencies"]
-
-[tool.setuptools.packages.find]
-exclude = ["instance"]
-
-
-[tool.setuptools.dynamic]
-dependencies = {file = ["requirements.txt"]}
-
-[build-system]
-requires = ["setuptools>=61.0"]
-build-backend = "setuptools.build_meta"
