+++ api/uwsgi.ini
@@ -1,5 +1,5 @@
 [uwsgi]
-module = flaskr:app
+module = flaskr:application
 master = 1
 
 ; gevent = 1000
@@ -9,6 +9,8 @@ socket = 0.0.0.0:3031
 http = 0.0.0.0:80
 stats = 0.0.0.0:9191
 stats-http = 1
+
+virtualenv=/opt/venv
 
 ; https=0.0.0.0:443,/crt/tls.crt,/crt/tls.key,HIGH
